requirements.txt
streamlit
gpogle-generativeai
python-dotenv
pdf2image